+++ kas/plugins/build.py
@@ -39,6 +39,7 @@ from kas.context import create_global_context
 from kas.config import Config
 from kas.libkas import find_program, run_cmd
 from kas.libcmds import Macro, Command
+from kas.libkas import setup_parser_common_args
 
 __license__ = 'MIT'
 __copyright__ = 'Copyright (c) Siemens AG, 2017-2018'
@@ -61,6 +62,7 @@ class Build:
             Setup the argument parser for the build plugin
         """
 
+        setup_parser_common_args(parser)
         parser.add_argument('extra_bitbake_args',
                             nargs='*',
                             help='Extra arguments to pass to bitbake')
