+++ kas/plugins/checkout.py
@@ -36,6 +36,7 @@
 from kas.context import create_global_context
 from kas.config import Config
 from kas.libcmds import Macro
+from kas.libkas import setup_parser_common_args
 
 __license__ = 'MIT'
 __copyright__ = 'Copyright (c) Siemens AG, 2017-2018'
@@ -50,7 +51,7 @@ class Checkout:
 
     @classmethod
     def setup_parser(cls, parser):
-        pass
+        setup_parser_common_args(parser)
 
     def run(self, args):
         ctx = create_global_context(args)
