+++ kas/plugins/shell.py
@@ -43,6 +43,7 @@ import sys
 from kas.context import create_global_context
 from kas.config import Config
 from kas.libcmds import Macro, Command
+from kas.libkas import setup_parser_common_args
 
 __license__ = 'MIT'
 __copyright__ = 'Copyright (c) Siemens AG, 2017-2018'
@@ -62,6 +63,7 @@ class Shell:
             Setup the argument parser for the shell plugin
         """
 
+        setup_parser_common_args(parser)
         parser.add_argument('-k', '--keep-config-unchanged',
                             help='Skip steps that change the configuration',
                             action='store_true')
